fix: harden patching, desktop streaming, and API security (39 review findings)
Address all findings from code review across agent, API, and viewer:

Critical: fix restorePointInfo struct layout (uint32 not int64), prevent
double-close panic in WebRTC session (sync.Once), remove UnlockOSThread
from wrong goroutine in MFT encoder, validate session ownership in
agentWs command relay.

High: add Chocolatey package name validation to prevent argument injection,
guard reboot manager Stop() against double-close, add delayMinutes bounds
check, default PatchAutoAcceptEula to false, fix PeerConnection leak on
viewer unmount, remove duplicate answer write path.

Medium: clamp video coordinates to prevent negatives, add SDP max size
validation (64KB), sync fullscreen state via event listener, add agent
offline warning in offer response, fix MFT incomplete output drain loop,
add NV12 buffer size verification, cap WebRTC bitrate at 20Mbps, fix
WUA countVar nil checks, add error status UI overlay, use native passive
wheel handler.

Low: replace WUA magic numbers with named constants, fix error wrapping,
optimize comCall for common arg counts, extract rebootStateToMap helper.

+++ agent/internal/patching/preflight_windows.go
@@ -326,18 +326,20 @@ func CreateRestorePoint(description string) error {
 		return fmt.Errorf("SRSetRestorePoint not available: %w", err)
 	}
 
-	// RESTOREPOINTINFOW structure
+	// RESTOREPOINTINFOW structure (must match Windows SDK RESTOREPOINTINFOW layout)
+	// SequenceNumber is DWORD (uint32), not int64. Using int64 would corrupt
+	// the Description field offset and cause SRSetRestorePoint to fail silently.
 	type restorePointInfo struct {
-		EventType         uint32
-		RestorePointType  uint32
-		SequenceNumber    int64
-		Description       [256]uint16
+		EventType        uint32
+		RestorePointType uint32
+		SequenceNumber   uint32
+		Description      [256]uint16
 	}
 
 	// STATEMGRSTATUS structure
 	type statemgrStatus struct {
 		Status         uint32
-		SequenceNumber int64
+		SequenceNumber uint32
 	}
 
 	const (
